# [wasm] Add WidenUpper and WidenLower SIMD intrins